fix(aiguard): resolve mypy errors from constants relocation
The refactor that moved Constant_Class to ddtrace.internal.constants and AI_GUARD
to ddtrace.aiguard._constants left 13 typing errors:

- Route the three Constant_Class importers to the canonical ddtrace.internal.constants
  (appsec._constants no longer explicitly re-exports it); this also removes a
  contrib -> appsec import in subprocess/constants.py.
- Annotate the relocated AIGuardConfig.reset and the new internal/aiguard/product.py
  module instead of relying on per-module mypy relaxations.
- Give AI_GUARD.ANOMALY_DETECTION_TAGS a concrete tuple[str, ...] type.
